moved last functions to Functions folder,moved data handling to actions.ts instead of state.ts
