+++ src/functions/audioMixer/actions.ts
@@ -1,11 +1,11 @@
 import { ActionId } from './actionId'
-import { getOptNumber } from '../../actions/index'
+import { getOptNumber } from '../../actions'
 import { ReqType, ActionType } from '../../enums'
-import { sendCommand } from '../../connection'
+import { sendCommand, GoStreamData } from '../../connection'
 import type { GoStreamInstance } from '../../index'
-import type { CompanionActionDefinitions } from '@companion-module/base'
 import { getChoices } from '../../choices'
 import { AudioMicChoices, AudioInputSourcesChoices, SwitchChoices } from '../../model'
+import type { CompanionActionDefinitions } from '@companion-module/base'
 
 export function create(instance: GoStreamInstance): CompanionActionDefinitions {
 	return {
@@ -300,4 +300,33 @@ export function create(instance: GoStreamInstance): CompanionActionDefinitions {
 			},
 		},
 	}
+}
+
+export function handleData(instance: GoStreamInstance, data: GoStreamData): boolean {
+	switch (data.id as ActionId) {
+		case ActionId.AudioTransition:
+			instance.states.AudioMixer.transitionEnabled = data.value[0] === 1 ? true : false
+			return true
+		case ActionId.AudioEnable: {
+			const audiotype = data.value[0]
+			const audiotypeValue = data.value[1]
+			if (audiotype == 0) {
+				instance.states.AudioMixer.micEnabled[0] = audiotypeValue
+			} else if (audiotype == 1) {
+				instance.states.AudioMixer.micEnabled[1] = audiotypeValue
+			} else if (audiotype == 2) {
+				instance.states.AudioMixer.inputEnabled[0] = audiotypeValue
+			} else if (audiotype == 3) {
+				instance.states.AudioMixer.inputEnabled[1] = audiotypeValue
+			} else if (audiotype == 4) {
+				instance.states.AudioMixer.inputEnabled[2] = audiotypeValue
+			} else if (audiotype == 5) {
+				instance.states.AudioMixer.inputEnabled[3] = audiotypeValue
+			} else if (audiotype == 6) {
+				instance.states.AudioMixer.auxEnabled[0] = audiotypeValue
+			}
+			return true
+		}
+	}
+	return false
 }
