+++ src/functions/downstreamKeyer/state.ts
@@ -1,8 +1,6 @@
 import { ActionId } from './actionId'
-import { sendCommand, GoStreamData } from '../../connection'
-import { ReqType, ActionType } from '../../enums'
-import { GoStreamInstance } from '../../index'
-import { getChoices } from './../../choices'
+import { sendCommand } from '../../connection'
+import { ReqType } from '../../enums'
 
 export type State = {
 	onAir: boolean
@@ -28,34 +26,6 @@ export function create(): DownstreamKeyerState {
 			shapedKey: false,
 		},
 	}
-}
-
-export function handleData(instance: GoStreamInstance, data: GoStreamData): boolean {
-	switch (data.id as ActionId) {
-		case ActionId.DskSourceFill: {
-			const select = getChoices(ActionType.DskSourceFill).find((s) => s.id === data.value[0])
-			if (select !== undefined) instance.states.DownstreamKeyer.fill = select
-			return true
-		}
-		case ActionId.DskSourceKey: {
-			const select = getChoices(ActionType.DskSourceFill).find((s) => s.id === data.value[0])
-			if (select !== undefined) instance.states.DownstreamKeyer.key = select
-			return true
-		}
-		case ActionId.DskControlInvert:
-			instance.states.DownstreamKeyer.invert = data.value[0] === 1 ? true : false
-			return true
-		case ActionId.DskMaskEnable:
-			instance.states.DownstreamKeyer.mask = data.value[0] === 1 ? true : false
-			return true
-		case ActionId.DskControlShapedKey:
-			instance.states.DownstreamKeyer.shapedKey = data.value[0] === 1 ? true : false
-			return true
-		case ActionId.DskOnAir:
-			instance.states.DownstreamKeyer.onAir = data.value[0] === 1 ? true : false
-			break
-	}
-	return false
 }
 
 export async function sync(): Promise<void> {
