+++ src/functions/live/state.ts
@@ -1,7 +1,6 @@
 import { ActionId } from './actionId'
-import { sendCommand, GoStreamData } from '../../connection'
+import { sendCommand } from '../../connection'
 import { ReqType } from '../../enums'
-import { GoStreamInstance } from '../../index'
 
 export type State = {
 	State: number
@@ -17,17 +16,6 @@ export function create(): LiveState {
 			State: 0,
 		},
 	}
-}
-
-export function handleData(instance: GoStreamInstance, data: GoStreamData): boolean {
-	switch (data.id as ActionId) {
-		case ActionId.Live: {
-			console.log('STATES', JSON.stringify(instance.states))
-			instance.states.Live.State = data.value[0]
-			return true
-		}
-	}
-	return false
 }
 
 export async function sync(): Promise<void> {
