+++ src/functions/settings/actions.ts
@@ -1,10 +1,11 @@
 import { ActionId } from './actionId'
-import { getOptNumber, getOptString } from './../../actions/index'
+import { getOptNumber, getOptString } from './../../actions'
 import { getChoices, SourcesToChoices } from './../../choices'
 import { ReqType, ActionType } from './../../enums'
-import { sendCommand } from './../../connection'
+import { sendCommand, GoStreamData } from './../../connection'
 import type { GoStreamInstance } from './../../index'
 import type { CompanionActionDefinitions } from '@companion-module/base'
+import { updateRecordVariables } from './../../variables'
 import {
 	SettingsAuxSourceChoices,
 	SettingsOutFormatChoices,
@@ -301,4 +302,55 @@ export function create(instance: GoStreamInstance): CompanionActionDefinitions {
 			},
 		},
 	}
+}
+export function handleData(instance: GoStreamInstance, data: GoStreamData): boolean {
+	switch (data.id as ActionId) {
+		case ActionId.AuxSource:
+			instance.states.SettingsProp.AuxSource = data.value[0]
+			return true
+		case ActionId.InputWindowLayout:
+			instance.states.SettingsProp.SettingsInputWindowLayout = data.value[0]
+			return true
+		case ActionId.MvMeter:
+			instance.states.SettingsProp.MvMeter[data.value[0]] = data.value[1]
+			return true
+		case ActionId.OutSource: {
+			const outType = data.value[0]
+			const outTypeValue = data.value[1]
+			const selectSource = getChoices(ActionType.SettingsoutSource).find((s) => s.id === outTypeValue)
+			if (outType === 0) {
+				if (selectSource !== undefined) {
+					instance.states.SettingsProp.OutSource.hdmi1 = selectSource
+				}
+			} else if (outType === 1) {
+				if (selectSource !== undefined) {
+					instance.states.SettingsProp.OutSource.hdmi2 = selectSource
+				}
+			} else if (outType === 2) {
+				if (selectSource !== undefined) {
+					instance.states.SettingsProp.OutSource.uvc = selectSource
+				}
+			}
+			return true
+		}
+		case ActionId.OutputColorSpace:
+			instance.states.SettingsProp.OutputColorSpace[data.value[0]] = data.value[1]
+			return true
+		case ActionId.OutFormat:
+			instance.states.SettingsProp.OutputFormat = data.value[0]
+			return true
+		case ActionId.MicInput:
+			instance.states.SettingsProp.MicInput[data.value[0]] = data.value[1]
+			return true
+		case ActionId.MvLayout:
+			instance.states.SettingsProp.MvLayout = data.value[0]
+			return true
+		case ActionId.SrcSelection:
+			instance.states.SettingsProp.SourceSelection[data.value[0]] = data.value[1]
+			return true
+		case ActionId.RecordTime:
+			updateRecordVariables(instance, data.value[0].toString())
+			return true
+	}
+	return false
 }
