+++ src/functions/superSource/state.ts
@@ -1,10 +1,6 @@
 import { ActionId } from './actionId'
-import { sendCommand, GoStreamData } from '../../connection'
-import { ActionType, ReqType } from '../../enums'
-import { GoStreamInstance } from '../../index'
-
-import { SuperSourceStyleChoices } from './../../model'
-import { getChoices } from './../../choices'
+import { sendCommand } from '../../connection'
+import { ReqType } from '../../enums'
 
 export type State = {
 	enable: boolean
@@ -36,73 +32,6 @@ export function create(): SuperSourceState {
 			},
 		},
 	}
-}
-
-export function handleData(instance: GoStreamInstance, data: GoStreamData): boolean {
-	switch (data.id as ActionId) {
-		case ActionId.SuperSourceBackground: {
-			const select = getChoices(ActionType.SuperSourceSource).find((s) => s.id === data.value[0])
-			if (select !== undefined) instance.states.SuperSource.background = select
-			return true
-		}
-		case ActionId.SuperSourceBorderBrightness: {
-			return true
-		}
-		case ActionId.SuperSourceBorderHue: {
-			return true
-		}
-		case ActionId.SuperSourceBorderSaturation: {
-			return true
-		}
-		case ActionId.SuperSourceBorderWidth: {
-			return true
-		}
-		case ActionId.SuperSourceControlStyle: {
-			const sschoice = SuperSourceStyleChoices.find((s) => s.id === data.value[0])
-			if (sschoice !== undefined) instance.states.SuperSource.controlStyle = sschoice
-			return true
-		}
-		case ActionId.SuperSourceControlYPosition: {
-			return true
-		}
-		case ActionId.SuperSourceEnable: {
-			instance.states.SuperSource.enable = data.value[0] == 1 ? true : false
-			return true
-		}
-		case ActionId.SuperSourceMaskEnable: {
-			const masktype = data.value[0]
-			const masktypeValue = data.value[1]
-			if (masktype === 0) {
-				instance.states.SuperSource.maskEnable.mask1 = masktypeValue === 1 ? true : false
-			} else {
-				instance.states.SuperSource.maskEnable.mask2 = masktypeValue === 1 ? true : false
-			}
-			return true
-		}
-		case ActionId.SuperSourceMaskHEnd: {
-			return true
-		}
-		case ActionId.SuperSourceMaskHStart: {
-			return true
-		}
-		case ActionId.SuperSourceMaskVEnd: {
-			return true
-		}
-		case ActionId.SuperSourceMaskVStart: {
-			return true
-		}
-		case ActionId.SuperSourceSource1: {
-			const select = getChoices(ActionType.SuperSourceSource).find((s) => s.id === data.value[0])
-			if (select !== undefined) instance.states.SuperSource.source1 = select
-			return true
-		}
-		case ActionId.SuperSourceSource2: {
-			const select = getChoices(ActionType.SuperSourceSource).find((s) => s.id === data.value[0])
-			if (select !== undefined) instance.states.SuperSource.source2 = select
-			return true
-		}
-	}
-	return false
 }
 
 export async function sync(): Promise<void> {
